+++ js-miniapp-bridge/src/types/custom-permissions.ts
@@ -2,7 +2,7 @@ export enum CustomPermissionName {
   USER_NAME = 'rakuten.miniapp.user.USER_NAME',
   PROFILE_PHOTO = 'rakuten.miniapp.user.PROFILE_PHOTO',
   CONTACT_LIST = 'rakuten.miniapp.user.CONTACT_LIST',
-  LOCATION = 'rakuten.miniapp.device.LOCATION'
+  LOCATION = 'rakuten.miniapp.device.LOCATION',
 }
 
 export enum CustomPermissionStatus {